+++ enginetest/queries/queries.go
@@ -1055,15 +1055,15 @@ Select * from (
 	{
 		Query: "SELECT pk DIV 2, SUM(c3) + sum(c3) as sum FROM one_pk GROUP BY 1 ORDER BY 1",
 		Expected: []sql.Row{
-			{int64(0), int64(28)},
-			{int64(1), int64(108)},
+			{int64(0), float64(28)},
+			{int64(1), float64(108)},
 		},
 	},
 	{
 		Query: "SELECT pk DIV 2, SUM(c3) + min(c3) as sum_and_min FROM one_pk GROUP BY 1 ORDER BY 1",
 		Expected: []sql.Row{
-			{int64(0), int64(16)},
-			{int64(1), int64(76)},
+			{int64(0), float64(16)},
+			{int64(1), float64(76)},
 		},
 		ExpectedColumns: sql.Schema{
 			{
@@ -1072,15 +1072,15 @@ Select * from (
 			},
 			{
 				Name: "sum_and_min",
-				Type: types.Int64,
+				Type: types.Float64,
 			},
 		},
 	},
 	{
 		Query: "SELECT pk DIV 2, SUM(`c3`) +    min( c3 ) FROM one_pk GROUP BY 1 ORDER BY 1",
 		Expected: []sql.Row{
-			{int64(0), int64(16)},
-			{int64(1), int64(76)},
+			{int64(0), float64(16)},
+			{int64(1), float64(76)},
 		},
 		ExpectedColumns: sql.Schema{
 			{
@@ -1089,7 +1089,7 @@ Select * from (
 			},
 			{
 				Name: "SUM(`c3`) +    min( c3 )",
-				Type: types.Int64,
+				Type: types.Float64,
 			},
 		},
 	},
@@ -4700,9 +4700,9 @@ Select * from (
 	{
 		Query: "SELECT SUM(i) + 1, i FROM mytable GROUP BY i ORDER BY i",
 		Expected: []sql.Row{
-			{int64(2), int64(1)},
-			{int64(3), int64(2)},
-			{int64(4), int64(3)},
+			{float64(2), int64(1)},
+			{float64(3), int64(2)},
+			{float64(4), int64(3)},
 		},
 	},
 	{
